pyproject.toml
[project]
name = "ai-comic"
version = "0.1.0"
description = "Local AI comic generator UI that drives a remote ComfyUI instance."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.0",
    "pillow>=10.0",
    "pydantic>=2.0",
    "sqlalchemy>=2.0",
    "websocket-client>=1.7",
    "requests>=2.31",
    "opencv-python-headless>=4.9",
    "numpy>=1.26",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config", "comfy", "generators", "library", "bubble", "ui"]

[tool.ruff]
line-length = 100
target-version = "py311"
